feat(replay): Do not capture replays < 5 seconds
Promotes the feature from https://github.com/getsentry/sentry-javascript/pull/7949 to GA.

Do not immediately flush on snapshot checkouts, instead delay by minimum flush delay (5 seconds). This means that we will not collect replays < 5 seconds. e.g. User opens site and immediately closes the tab.

Closes https://github.com/getsentry/team-replay/issues/65
